refactor: move twitter file and improve client
